Feature ETP-4034: Extract shared components to fix Sonar CPD violations
