+++ tools/app-shell/src/windows/custom/return-to-vendor-shipment/useReturnToVendorPdf.js
@@ -1,7 +1,13 @@
 import { useUI } from '@/i18n';
 import {
-  COMMON_HANDLEBARS_HELPERS,
+  RETURN_DOC_HELPERS,
   COMMON_PDF_CSS,
+  MOVEMENT_TEMPLATE_OPEN,
+  MOVEMENT_TEMPLATE_HEADER,
+  MOVEMENT_TEMPLATE_PARTIES,
+  MOVEMENT_TEMPLATE_SIGNATURE,
+  MOVEMENT_TEMPLATE_NOTES,
+  MOVEMENT_TEMPLATE_FOOTER,
   fetchJson,
   fetchAll,
   fetchOptionalJson,
@@ -12,51 +18,10 @@ import {
   usePdfGenerator,
 } from '../shared/pdfUtils.js';
 
-const HELPERS = `
-function fmt(v) {
-  if (v == null || v === '') return '0';
-  var n = Number(v);
-  if (isNaN(n)) return String(v);
-  return new Intl.NumberFormat('es', { minimumFractionDigits: 0, maximumFractionDigits: 3 }).format(n);
-}
-` + COMMON_HANDLEBARS_HELPERS;
-
-
-const TEMPLATE = `<!DOCTYPE html>
-<html lang="es"><head><meta charset="UTF-8"><style>{{{css}}}</style></head>
-<body>
-<div class="doc">
-
-  <div class="doc-header">
-    <div class="doc-header-left">
-      {{#if companyLogoDataUrl}}<img class="doc-logo-img" src="{{companyLogoDataUrl}}" alt="">{{/if}}
-      <div class="doc-header-company">
-        <div class="company-name">{{companyName}}</div>
-        {{#if companyTaxId}}<div class="company-cif">{{labels.taxId}} {{companyTaxId}}</div>{{/if}}
-      </div>
-    </div>
-    <div class="doc-header-meta">
-      <div class="doc-label">{{labels.title}}</div>
-      <div class="doc-num">{{documentNo}}</div>
-    </div>
-  </div>
-
-  <div class="doc-parties">
-    <div class="doc-party-card">
-      <div class="eyebrow">{{labels.issuerSection}}</div>
-      <div class="party-name">{{companyName}}</div>
-      {{#if companyAddress1}}<div class="party-line">{{companyAddress1}}</div>{{/if}}
-      {{#if companyAddress2}}<div class="party-line">{{companyAddress2}}</div>{{/if}}
-      {{#if companyCityLine}}<div class="party-line">{{companyCityLine}}</div>{{/if}}
-      {{#if companyTaxId}}<div class="party-line muted">{{labels.taxId}} {{companyTaxId}}</div>{{/if}}
-    </div>
-    <div class="doc-party-card">
-      <div class="eyebrow">{{labels.vendorSection}}</div>
-      <div class="party-name">{{vendorName}}</div>
-      {{#each vendorAddressLines}}<div class="party-line">{{this}}</div>{{/each}}
-    </div>
-  </div>
-
+const TEMPLATE = MOVEMENT_TEMPLATE_OPEN
++ MOVEMENT_TEMPLATE_HEADER
++ MOVEMENT_TEMPLATE_PARTIES
++ `
   <div class="doc-movement">
     <div class="doc-movement-cell">
       <div class="cell-label">{{labels.date}}</div>
@@ -91,33 +56,10 @@ const TEMPLATE = `<!DOCTYPE html>
       </tr>
       {{/each}}
     </tbody>
-  </table>
-
-  <div class="doc-signature">
-    <div class="doc-sig-field">
-      <span class="sig-label">{{labels.signatureIssuer}}</span>
-      <div class="sig-line"></div>
-    </div>
-    <div class="doc-sig-field">
-      <span class="sig-label">{{labels.signatureDate}}</span>
-      <div class="sig-line"></div>
-    </div>
-  </div>
-
-  {{#if notes}}
-  <div class="doc-observ">
-    <div class="eyebrow">{{labels.notes}}</div>
-    <div class="body">{{notes}}</div>
-  </div>
-  {{/if}}
-
-  <div class="doc-footer">
-    <span>{{companyName}}</span>
-    <span>{{labels.page}} 1</span>
-  </div>
-
-</div>
-</body></html>`;
+  </table>`
++ MOVEMENT_TEMPLATE_SIGNATURE
++ MOVEMENT_TEMPLATE_NOTES
++ MOVEMENT_TEMPLATE_FOOTER;
 
 async function buildReturnToVendorData(shipmentId, base, token) {
   const [header, linesRaw, session] = await Promise.all([
@@ -169,14 +111,14 @@ export function useReturnToVendorPdf(shipmentId, apiBaseUrl, token) {
   const ui = useUI();
   return usePdfGenerator(shipmentId, apiBaseUrl, token, (id, base, tok) => {
     const labels = getReturnToVendorPdfLabels(ui);
-    return buildReturnToVendorData(id, base, tok).then((data) => renderPdf(TEMPLATE, COMMON_PDF_CSS, HELPERS, { ...data, labels }));
+    return buildReturnToVendorData(id, base, tok).then((data) => renderPdf(TEMPLATE, COMMON_PDF_CSS, RETURN_DOC_HELPERS, { ...data, labels }));
   });
 }
 
 export async function generateReturnToVendorPdf(shipmentId, apiBaseUrl, token, labels) {
   const base = apiBaseUrl.replace(/\/[^/]+$/, '');
   const data = await buildReturnToVendorData(shipmentId, base, token);
-  return renderPdf(TEMPLATE, COMMON_PDF_CSS, HELPERS, { ...data, labels });
+  return renderPdf(TEMPLATE, COMMON_PDF_CSS, RETURN_DOC_HELPERS, { ...data, labels });
 }
 
 export function getReturnToVendorPdfLabels(ui) {
